doc: add known-issues to nrf build
Run known-issues script from Zephyr on nrf build to generate a clean
doc.warnings file. The current issues are due to a limitation in Sphinx
which does not allow multiple symbols with the same name, eg, struct
and function using the same name. The Sphinx fix is already
work-in-progress, so this change should be dropped eventually.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>